Allow kmacro-x-mc-atomic-undo-mode to also undo the original change
Note: Usually any such change-group operations are wrapped in
unwind-protect.  I don't think this is necessary here, as we use the
minor-mode (de)activation logic to first activate the change-group on
mode enabling and then to accept-change-group it on mode disabling.
No errors in-between them should affect the fact we'll disable the
mode at some point and the cleanup will happen.  This is different
from doing it within a single function which may be interrupted with
an error.

https://www.gnu.org/software/emacs/manual/html_node/elisp/Atomic-Changes.html